aforementioned new code

+++ cluster.py
@@ -5,7 +5,7 @@ import os
 import shutil
 
 from openff.qcsubmit.results import OptimizationResultCollection
-from openff.toolkit import ForceField
+from openff.toolkit import Molecule
 from tqdm import tqdm
 from vflib.draw import draw_rdkit
 
@@ -21,46 +21,60 @@ def find(lst, fun):
     return None
 
 
-sage = ForceField("openff-2.1.0.offxml")
-opt = OptimizationResultCollection.parse_file("filtered-opt.json")
-molecules = opt.to_molecules()
-
-with open("bonds.dat", "r") as inp:
-    for line in inp:
-        if line.startswith("#"):
+def deduplicate_by(lst, fn):
+    "Deduplicate `lst` by the key generated from `fn`."
+    keys = set()
+    ret = []
+    for elt in lst:
+        key = fn(elt)
+        if key in keys:
             continue
-        # take just the first data line for now
-        break
+        keys.add(key)
+        ret.append(elt)
+    return ret
 
-sp = line.split()
-smirks = sp[0]
 
-out_dir = "bond0"
-shutil.rmtree(out_dir)
-os.mkdir(out_dir)
+def load_smirks(filename):
+    """Extracts the initial SMIRKS pattern from lines of the form below.
 
-seen = set()
-for m, molecule in tqdm(
-    enumerate(molecules), total=len(molecules), desc="Labeling molecules"
-):
-    matches = molecule.chemical_environment_matches(smirks)
-    if matches:
-        _, esp = espaloma_label(molecule)
-        esp = [
-            (b.atom1 - 1, b.atom2 - 1)
-            for b in esp["bonds"]
-            if (b.atom1 - 1, b.atom2 - 1) in matches and b.k < 720.0
-        ]
-        if esp:
-            inchi = molecule.to_inchikey()
-            if inchi not in seen:
+    [#6X4:1]-[#1:2] 50377  719.64  740.36  740.36  740.36
+    """
+    with open(filename, "r") as inp:
+        return [line.split()[0] for line in inp if not line.startswith("#")]
+
+
+opt = OptimizationResultCollection.parse_file("filtered-opt.json")
+
+# demand execution here so I can deduplicate and reuse molecules
+molecules = [m for m in tqdm(opt.to_molecules(), desc="Converting molecules")]
+print(len(molecules), " initially")
+molecules = deduplicate_by(molecules, Molecule.to_inchikey)
+print(len(molecules), " after dedup")
+
+smirkss = load_smirks("bonds.dat")
+for s, smirks in tqdm(enumerate(smirkss), desc="Processing smirks"):
+    out_dir = f"bond{s:02d}"
+    if os.path.exists(out_dir):
+        shutil.rmtree(out_dir)
+    os.mkdir(out_dir)
+
+    for m, molecule in enumerate(molecules):
+        matches = molecule.chemical_environment_matches(smirks)
+        if matches:
+            _, esp = espaloma_label(molecule)
+            esp = [
+                (b.atom1 - 1, b.atom2 - 1)
+                for b in esp["bonds"]
+                if (b.atom1 - 1, b.atom2 - 1) in matches and b.k < 720.0
+            ]
+            if esp:
                 draw_rdkit(
                     molecule,
                     smirks,
                     matches=esp,
                     filename=f"{out_dir}/mol{m:04d}.png",
                 )
-                seen.add(inchi)
+    break
 
 
 # for bond0 it's already looking like the much lower espaloma values are for a
